moved ServiceClient code to own file
